Add plugin to read pickle files from ESPP v1 database file
Selling of shares isn't verified, but seems at least somewhat
functional.

Note: A lot of information is missing from the pickle file,
and therefore missing from the transaction file - like dates
of purchase and deposit.

The pickle-file must be specified as the transaction file input
to the 'espp2_transnorm' tool, and the resulting transactions
will be stored in the output-file in *.json format.

The format must be specified as 'pickle'.