morgan-import: Improve import-function by requiring two files
Require one file with full history in HTML format for everything
up until 2021-12-31, and another file containing everything
from 2022-01-01 to 2022-12-31.

The first file is needed to establish the opening balance for 2022,
and the second file is needed to calculate the tax and update the
balance with what happened in 2022.

This commit has issues, e.g. skjermingsfradrag is too high, and
the opening balance for cash is zero.

This needs more work.