transactions: Add CASHADJUST transaction for ad-hoc cash updates
This is a flexible transaction entry type for modifying the cash
holdings by a positive or negative delta (debit or credit),

The description given in the transaction will show up in the
cash-ledger and holdings.

The initial use for this will be to bootstrap the cash-holdings
when importing from Morgan Stanley.