morgan-import: Add arg force_tax_deduct_reset to do_holdings_1()
This argument (Default False) must be used when importing the
morgan history to create the initial holdings for 2021.

This argument, when True, will cause the shielding for 2021 to
be calculated conservatively, under the assumption that shielding
available for tax-reduction in 2021 was claimed, and thus can't
be carried over into 2022 through the holdings file for 2021.

Going forward, the holding file will be used as a basis for next
years tax-calculation, so the need for this function only applies
to those that need to bootstrap old history from Morgan Stanley.