morgan-import: Use an auxiliary dict() for importer to provide info
The Morgan Stanley importer bootstraps holdings for 2021 by means
of the holding tables in the HTML files with statements.

The "skjermingsfradrag" needs to be calculated differently for this
case, since the complete history is not available, and we must make
some assumptions and conservative choices for 2021/2022 in order
to not claim too much tax deduction for skjermingsfradrag.

Communicate this need through the auxiliary import data, so the
importer chosen (and specific circumstances), can guide the
correcting of the skjermingsfradrag later in the process, when
the holdings have been established.