morgan-import: Let elem_enter() search for the element to enter
There has been cases where e.g. a 'table' tag is followed, not
by 'tbody' as expected, but rather a comment, and then the 'tbody'.

By letting the elem_enter() function search forward for the right
tag, and not bail out if the next tag is not exactly right, is a
simple way to skip the comment and make it work.

There may be cases where this is not desirable, but we'll deal
with those when they happen.