morgan-import: Improve opening balace parser for shares
This is risky, so keep it disabled for now. This seems to kick in
when trying to parse older years as transactions, and it is
unclear if the opening balance ever has a matching buy transaction.

For this reason, we don't enable it yet, to gain more experience
with files that needs to parse such records. The normal use-case
for the tool (to extract transactions only from the previous year)
does not seem to require this functionality.