Don't require transactions outside of tax-year for morgan
Since ESSP purchase is listed at e.g. 2024-12-31, and not some
days into the new year (as for Schwab), we don't need any
transactions in the following year to get ESPP purchases correct.

So don't require the import file to contain any transactions
from the year following the tax-year in question, for morgan.