morgan: Use 'Entry Date' for date of sale for shares
This date is listed as "Settlement date" in the withdrawals-tables,
but has previously not been used for date of sale. This has
probably been done to get dividend verification to pass, but it
is unclear why this change needs to be reversed now.

This change causes our test-files for 2024 to pass, but this might
not be the correct fix depending on what the (current) settlement
date really means.