morgan: Extract discounted purchase price from ESPP holdings table
It looks like the "Purchase Price" column in the ESPP Holdings
table contains the discounted purchase price. At least this seems
to be the case for older purchases, but in December 2024 it looks
like the purchase price is the same as the fair market value.

Extract the numbers anyway, in hopes that this will be more useful
than not extracting these numbers. They are used to provide
numbers on the gains from ESPP purchases, which includes the
discounted price originally paid.