Fix crash when fee's field on a sale is empty
If a sale has no fees the fee's field is empty, and causes a crash.
Fixed the issue and added error logging to print what we are trying to
parse, as well as raise on the issue as rasing on the assert does not
provide much of context around the issue.