morgan: Use withdrawal gross proceeds as accurate sales value
An actual sale may be different from the per share price given
with two decimal places in the transaction listing. Using the
quantity and price from this transaction listing may cause
slightly incorrect calculation of the gross amount.

Try to work around this problem by extracting the gross proceeds
value from the Withdrawal tables, and use this value instead
for the sale (rather than the product of qty and price).

Identify what gross proceed values to use instead by looking for
withdrawal tables for the date in question, and where the value
is sufficiently close to be explained by a smaller than a cent
delta to the share price.